feat: add DTO to the other end points

+++ CRUD-Game-List/DTOs/CategoryDtos.cs
@@ -2,7 +2,7 @@
 {
 	public record CategoryCreateDto(string Name);
 
-	public record CategoryUpdateDto(int Id, string Name);
+	public record CategoryUpdateDto(long Id, string Name);
 
-	public record CategoryDto(int Id, string Name);
+	public record CategoryDto(long Id, string Name);
 }
